requirements.txt
# project dependencies
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
asyncpg>=0.27.0
pydantic>=1.10.0
requests>=2.25.0
python-dotenv>=0.19.0
sqlalchemy[asyncio]>=1.4.0
alembic>=1.8.0
